requirements.txt
fastapi
uvicorn
pandas
openpyxl
requests
beautifulsoup4
psycopg2-binary
openai
python-dotenv
sqlalchemy
httpx
pgvector
